+++ src/test/java/io/sentry/unit/autoinstall/graphql/Graphql22AutoInstallTest.kt
@@ -1,7 +1,7 @@
 package io.sentry.unit.autoinstall.graphql
 
 import io.sentry.autoinstall.AutoInstallState
-import io.sentry.autoinstall.graphql.GraphqlInstallStrategy
+import io.sentry.autoinstall.graphql.Graphql22InstallStrategy
 import io.sentry.unit.fakes.CapturingTestLogger
 import org.apache.maven.model.Dependency
 import org.eclipse.aether.artifact.Artifact
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test
 import org.slf4j.Logger
 import kotlin.test.assertTrue
 
-class GraphqlAutoInstallTest {
+class Graphql22AutoInstallTest {
     class Fixture {
         val logger = CapturingTestLogger()
         val dependencies = ArrayList<Dependency>()
@@ -19,9 +19,9 @@ class GraphqlAutoInstallTest {
 
         fun getSut(
             installGraphql: Boolean = true,
-            graphqlVersion: String = "2.0.0",
-        ): GraphqlInstallStrategy {
-            installState = AutoInstallState("6.25.2")
+            graphqlVersion: String = "22.0",
+        ): Graphql22InstallStrategy {
+            installState = AutoInstallState("8.0.0")
             installState.isInstallGraphql = installGraphql
 
             resolvedArtifacts.add(
@@ -33,7 +33,7 @@ class GraphqlAutoInstallTest {
                 ),
             )
 
-            return GraphqlInstallStrategyImpl(logger)
+            return Graphql22InstallStrategyImpl(logger)
         }
     }
 
@@ -47,10 +47,10 @@ class GraphqlAutoInstallTest {
 
         assertTrue {
             fixture.logger.capturedMessage ==
-                "sentry-graphql won't be installed because it was already installed directly"
+                "sentry-graphql-22 won't be installed because it was already installed directly"
         }
 
-        assertTrue(fixture.dependencies.none { it.groupId == "io.sentry" && it.artifactId == "sentry-graphql" })
+        assertTrue(fixture.dependencies.none { it.groupId == "io.sentry" && it.artifactId == "sentry-graphql-22" })
     }
 
     @Test
@@ -61,26 +61,26 @@ class GraphqlAutoInstallTest {
 
         assertTrue {
             fixture.logger.capturedMessage ==
-                "sentry-graphql was successfully installed with version: 6.25.2"
+                "sentry-graphql-22 was successfully installed with version: 8.0.0"
         }
 
-        assertTrue(fixture.dependencies.any { it.groupId == "io.sentry" && it.artifactId == "sentry-graphql" })
+        assertTrue(fixture.dependencies.any { it.groupId == "io.sentry" && it.artifactId == "sentry-graphql-22" })
     }
 
     @Test
-    fun `when graphql version is too high logs a message and does nothing`() {
-        val sut = fixture.getSut(graphqlVersion = "22.1")
+    fun `when graphql version is too low logs a message and does nothing`() {
+        val sut = fixture.getSut(graphqlVersion = "21.9")
 
         sut.install(fixture.dependencies, fixture.resolvedArtifacts, fixture.installState)
 
         assertTrue {
             fixture.logger.capturedMessage ==
-                "sentry-graphql won't be installed because the current " +
-                "version is higher than the maximum supported version 21.9999.9999"
+                "sentry-graphql-22 won't be installed because the current " +
+                "version is lower than the minimum supported version 22.0.0"
         }
 
-        assertTrue(fixture.dependencies.none { it.artifactId == "sentry-graphql" })
+        assertTrue(fixture.dependencies.none { it.artifactId == "sentry-graphql-22" })
     }
 
-    private class GraphqlInstallStrategyImpl(logger: Logger) : GraphqlInstallStrategy(logger)
+    private class Graphql22InstallStrategyImpl(logger: Logger) : Graphql22InstallStrategy(logger)
 }